Money types: convert Sales domain Double -> BigDecimal (slice 5/N)

+++ billbull-backend/src/main/java/com/billbull/backend/sales/invoice/PriceHistoryDTO.java
@@ -1,10 +1,11 @@
 package com.billbull.backend.sales.invoice;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record PriceHistoryDTO(
         LocalDate date,
         String invoiceNo,
         String customerName,
-        Double rate) {
+        BigDecimal rate) {
 }
